se a agregado soporte para tasas flotantes (alfa beta desercion)

+++ config_loader.py
@@ -13,6 +13,6 @@ def load_config(): #leo el archivo y estraigo los numeritos y despues los retorn
     file_flag = open(config_location,'r')
     
     for x in file_flag.readlines():
-        data.append([int(i) for i in x.split()])
+        data.append([float(i) for i in x.split()])
     return data
